# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PowerPulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  phone?: string | null;
  photo?: string | null;
  reputation: number;
  createdAt: string;
}

export interface UpdateUserProfileRequest {
  name?: string;
  phone?: string;
  photo?: string;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  outage: 'outage',
  restoration: 'restoration',
  transformer: 'transformer',
} as const;

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export interface Report {
  id: number;
  userId: string;
  userName?: string | null;
  type: ReportType;
  status: ReportStatus;
  latitude: number;
  longitude: number;
  address?: string | null;
  description?: string | null;
  faultType?: string | null;
  confirmations: number;
  disputes: number;
  confidence: number;
  verified: boolean;
  createdAt: string;
  updatedAt: string;
  resolvedAt?: string | null;
}

export type CreateReportRequestType = typeof CreateReportRequestType[keyof typeof CreateReportRequestType];


export const CreateReportRequestType = {
  outage: 'outage',
  restoration: 'restoration',
  transformer: 'transformer',
} as const;

export interface CreateReportRequest {
  type: CreateReportRequestType;
  latitude: number;
  longitude: number;
  address?: string;
  description?: string;
  faultType?: string;
}

export type UpdateReportRequestStatus = typeof UpdateReportRequestStatus[keyof typeof UpdateReportRequestStatus];


export const UpdateReportRequestStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export interface UpdateReportRequest {
  status?: UpdateReportRequestStatus;
  description?: string;
  resolvedAt?: string;
}

export type VerifyReportRequestVote = typeof VerifyReportRequestVote[keyof typeof VerifyReportRequestVote];


export const VerifyReportRequestVote = {
  confirm: 'confirm',
  dispute: 'dispute',
} as const;

export interface VerifyReportRequest {
  vote: VerifyReportRequestVote;
}

export interface Notification {
  id: number;
  userId: string;
  title: string;
  body: string;
  read: boolean;
  createdAt: string;
}

export type StatisticsCurrentStatus = typeof StatisticsCurrentStatus[keyof typeof StatisticsCurrentStatus];


export const StatisticsCurrentStatus = {
  available: 'available',
  outage: 'outage',
  unstable: 'unstable',
  unknown: 'unknown',
} as const;

export interface Statistics {
  reliabilityPercent: number;
  avgOutageDurationMinutes: number;
  monthlyOutages: number;
  longestOutageMinutes: number;
  currentStatus: StatisticsCurrentStatus;
  activeReports: number;
  totalReports: number;
}

export interface SavedLocation {
  id: number;
  userId: string;
  name: string;
  latitude: number;
  longitude: number;
  address?: string | null;
  createdAt: string;
}

export interface CreateSavedLocationRequest {
  name: string;
  latitude: number;
  longitude: number;
  address?: string;
}

export type ListReportsParams = {
lat?: number;
lng?: number;
/**
 * Radius in km
 */
radius?: number;
type?: ListReportsType;
status?: ListReportsStatus;
limit?: number;
};

export type ListReportsType = typeof ListReportsType[keyof typeof ListReportsType];


export const ListReportsType = {
  outage: 'outage',
  restoration: 'restoration',
  transformer: 'transformer',
} as const;

export type ListReportsStatus = typeof ListReportsStatus[keyof typeof ListReportsStatus];


export const ListReportsStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export type GetStatisticsParams = {
lat?: number;
lng?: number;
};