# lib/api-zod/src/generated/types/createSavedLocationRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PowerPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSavedLocationRequest {
  name: string;
  latitude: number;
  longitude: number;
  address?: string;
}