# lib/api-zod/src/generated/types/updateUserProfileRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PowerPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUserProfileRequest {
  name?: string;
  phone?: string;
  photo?: string;
}