# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PowerPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  phone?: string | null;
  photo?: string | null;
  reputation: number;
  createdAt: Date;
}