# lib/api-zod/src/generated/types/verifyReportRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PowerPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerifyReportRequestVote } from './verifyReportRequestVote';

export interface VerifyReportRequest {
  vote: VerifyReportRequestVote;
}